install_pmbootstrap.sh: only show output on error
It does not make sense to use -q and --details to stdout. Remove -q and
redirect the output to another file, so we even have the log if
pmbootstrap fails before the workdir and logfile are created. Display it
if something goes wrong. While at it, update the message printed on
error.